\name{BRAINscape}
\alias{BRAINscape}
\docType{package}
\title{
\packageTitle{BRAINscape}
}
\description{
\code{BRAINscape} provides an interactive Shiny application for visualizing results from eQTL, DESeq2, and GWAS analyses performed on the MU-BRAIN dataset. It supports both combined and stratified analyses across Hispanic and Non-Hispanic White (NHW) brain samples, enabling users to explore population-specific genetic and transcriptomic insights.
}

\usage{
library(BRAINscape)
BRAINscape::launchApp()
}

\references{
   Yang, Z., Cieza, B., Reyes-Dumeyer, D., Lee, A., Ma, Y., Yilmaz, E., Lantigua, R., Miller, G.W., Brown, L.M., Honig, L. and Ciener, B., 2024. MU-BRAIN: MUltiethnic Brain Rna-seq for Alzheimer INitiative. bioRxiv, pp.2024-02.
}

\keyword{Genetics}
\keyword{eQTL}
\keyword{GWAS}
\keyword{Transcriptomics}
\keyword{Colocalization}
\keyword{Alzheimer's Disease}

%% Uncomment below to imitate parts of library(help = BRAINscape)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{BRAINscape}}
%\section{Documentation Index}{\packageIndices{BRAINscape}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
