% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_eqtl}
\alias{load_eqtl}
\title{Load eQTL data}
\usage{
load_eqtl(path, ethnicity, type)
}
\arguments{
\item{path}{The base path to data files}

\item{ethnicity}{"Hispanics" or "NonHispanics"}

\item{type}{"cis" or "trans"}
}
\value{
The loaded eQTL data
}
\description{
Load eQTL data
}
