\name{BRAINscape}
\alias{BRAINscape}
\docType{package}
\title{
\packageTitle{BRAINscape}
}
\description{
\code{BRAINscape} allows user to explore the Expression Quantitative Trait Loci (eQTL) and Genome-wide Association Study (GWAS) analysis performed on Hispanics and Non-Hispanics Whites (NHW) individuals in MU-BRAIN dataset.
}
\usage{
library(BRAINscape)
BRAINscape::launchApp()
}

\references{
   Yang, Z., Cieza, B., Reyes-Dumeyer, D., Lee, A., Ma, Y., Yilmaz, E., Lantigua, R., Miller, G.W., Brown, L.M., Honig, L. and Ciener, B., 2024. MU-BRAIN: MUltiethnic Brain Rna-seq for Alzheimer INitiative. bioRxiv, pp.2024-02.
}

\keyword{Genetics}
\keyword{eQTL}
\keyword{GWAS}
\keyword{Alzheimer's Disease}

%% Uncomment below to imitate parts of library(help = BRAINscape)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{BRAINscape}}
%\section{Documentation Index}{\packageIndices{BRAINscape}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
